Fixed images
The images.py script now uses regex to find and replace images in
markdown files. This allows you to keep the description of the image
instead of it replacing with "Image Description".

The images.py script also takes an argument from the bash script now
which allows the bash script to pass through the type of post being
updated so the images.py script can dynamically update the file path.